Refactor multiplexer to be an interface
